fix: resize the Target with the Image
Change to how the ViaDataset was working to enable transforms to be applied to both the picture and the target.

The key here is the 'ViaTransform', which actually gets two args, the image and the target.

Any common transform can be applied to it by using the to_via_transform() decorator.

The composition has to be done with ComposeViaTransform.